requirements.txt
google-analytics-admin
google-cloud-bigquery